#4 : Listing page (Loic + Daniel + Kayne)
Added realtime database to home fragment
